Add the draggable behaviour to AiNimationDragToReplay Composable for replaying animations
